// src/app/enums/job-type.ts
export enum JobType {
    S = 'sécurité et protection',
    P = 'observation et surveillance',
    E = 'travaux physiques le soldat d\'infanterie',
    C = 'management et communication',
    I = 'science & recherche',
    A = 'Maître d\'armes ou spécialiste en arts martiaux',
    L = 'explorateur et aventurier'
}